Moved config into client_core
