+++ clients/client-core/src/config/persistence/key_pathfinder.rs
