+++ clients/client-core/src/config/template.rs
