+++ clients/native/src/lib.rs
@@ -14,5 +14,4 @@
 
 pub mod built_info;
 pub mod client;
-pub mod config;
 pub mod websocket;
