+++ clients/native/src/main.rs
@@ -16,8 +16,7 @@ use clap::{App, ArgMatches};
 
 pub mod built_info;
 pub mod client;
-mod commands;
-pub mod config;
+pub mod commands;
 pub mod websocket;
 
 fn main() {
